Extract shared log summarizer for client and API loggers
Move redaction/truncation logic into src/shared/logSummarize.ts with
environment-specific options. logger.ts and api/_utils/_logging.ts become
thin wrappers preserving existing summarizeForLog/summarizeForApiLog APIs.
